Adding to the main view when login is success or registration is done
