+++ package.json
@@ -18,7 +18,8 @@
     "jquery": "3.4.1",
     "popper.js": "1.16.0",
     "react": "16.12.0",
-    "react-dom": "16.12.0"
+    "react-dom": "16.12.0",
+    "react-router-dom": "^5.1.2"
   },
   "devDependencies": {
     "@babel/core": "7.7.2",
