+++ src/components/main.js
@@ -1,0 +1,6 @@
+import React from "react";
+
+export const Main = props => {
+  const { email } = props.history.location.state;
+  return <div>Hello {email}</div>;
+};
